requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
tortoise-orm==0.21.6
asyncpg==0.29.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.3.0
aiohttp==3.10.5
aerich==0.7.2

